+++ src/App.js
@@ -1,7 +1,7 @@
 import './App.css';
 import { useEffect } from 'react';
 import { useDispatch } from 'react-redux';
-import { regions } from './api/codes';
+import { regions } from './redux/co2Emissions/api/codes';
 import Regions from './components/regions';
 import { getCo2DataApi } from './redux/co2Emissions/co2Emissions';
 
@@ -10,11 +10,11 @@ function App() {
 
   useEffect(() => {
     dispatch(getCo2DataApi(regions));
-}, []);
+  }, []);
 
   return (
     <div className="App">
-        <Regions />
+      <Regions />
     </div>
   );
 }
