+++ src/components/regions.js
@@ -1,13 +1,13 @@
-import React, { useEffect } from "react";
-import { useStore } from "react-redux";
+import React, { useState } from 'react';
+import { useStore } from 'react-redux';
 
 const Regions = () => {
-    const regions = useStore()
-    return (
-        <div> 
-        <p>{regions[1].value}</p>
-        </div>
-    )
-}
+  const regions = useState();
+  return (
+    <div>
+      <p>{regions[1].value}</p>
+    </div>
+  );
+};
 
 export default Regions;
