+++ src/redux/co2Emissions/api/codes.js
@@ -1,0 +1,216 @@
+export const regions = 'eas;ecs;lcn;mea;nac;sas;ssf';
+
+export const ecs = ['AL',
+  'AD',
+  'AM',
+  'AT',
+  'AZ',
+  'BE',
+  'BA',
+  'HR',
+  'CY',
+  'DK',
+  'EE',
+  'FO',
+  'FI',
+  'FR',
+  'GE',
+  'DE',
+  'GI',
+  'GR',
+  'GL',
+  'VA',
+  'IS',
+  'IE',
+  'IM',
+  'IT',
+  'KZ',
+  'KG',
+  'LV',
+  'LI',
+  'LT',
+  'LU',
+  'MT',
+  'MC',
+  'ME',
+  'NL',
+  'MK',
+  'NO',
+  'PT',
+  'SM',
+  'RS',
+  'SI',
+  'ES',
+  'SE',
+  'CH',
+  'TJ',
+  'TR',
+  'TM',
+  'GB',
+  'UZ'].join(';');
+
+export const eas = ['AS',
+  'AU',
+  'BN',
+  'KH',
+  'CN',
+  'FJ',
+  'PF',
+  'GU',
+  'HK',
+  'ID',
+  'JP',
+  'KI',
+  'KP',
+  'KR',
+  'LA',
+  'MO',
+  'MY',
+  'MH',
+  'FM',
+  'MN',
+  'MM',
+  'NR',
+  'NC',
+  'NZ',
+  'MP',
+  'PW',
+  'PG',
+  'PH',
+  'WS',
+  'SG',
+  'SB',
+  'TW',
+  'TH',
+  'TL',
+  'TO',
+  'TV',
+  'UM',
+  'VU',
+  'VN'].join(';');
+
+export const lcn = ['AG',
+  'AR',
+  'AW',
+  'BS',
+  'BB',
+  'BZ',
+  'BO',
+  'BR',
+  'KY',
+  'CL',
+  'CO',
+  'CR',
+  'CU',
+  'CW',
+  'DM',
+  'DO',
+  'EC',
+  'SV',
+  'GD',
+  'GT',
+  'GY',
+  'HT',
+  'HN',
+  'JM',
+  'MX',
+  'NI',
+  'PA',
+  'PY',
+  'PE',
+  'PR',
+  'KN',
+  'LC',
+  'MF',
+  'VC',
+  'SX',
+  'SR',
+  'TT',
+  'TC',
+  'UY',
+  'VE',
+  'VG',
+  'VI'].join(';');
+
+export const mea = ['DZ',
+  'BH',
+  'EG',
+  'IQ',
+  'IL',
+  'JO',
+  'KW',
+  'LB',
+  'LY',
+  'MA',
+  'OM',
+  'PS',
+  'QA',
+  'SA',
+  'SD',
+  'SY',
+  'TN',
+  'AE',
+  'YE'].join(';');
+
+export const nac = ['BM',
+  'CA',
+  'US'].join(';');
+
+export const sas = ['AF',
+  'BD',
+  'BT',
+  'IN',
+  'IR',
+  'MV',
+  'NP',
+  'PK',
+  'LK'].join(';');
+
+export const ssf = ['AO',
+  'BJ',
+  'BF',
+  'BI',
+  'CV',
+  'CM',
+  'CF',
+  'TD',
+  'KM',
+  'CG',
+  'CD',
+  'CI',
+  'DJ',
+  'GQ',
+  'ER',
+  'SZ',
+  'ET',
+  'GA',
+  'GM',
+  'GH',
+  'GN',
+  'GW',
+  'KE',
+  'LS',
+  'LR',
+  'MG',
+  'MW',
+  'ML',
+  'MR',
+  'MU',
+  'YT',
+  'MZ',
+  'NA',
+  'NE',
+  'NG',
+  'RW',
+  'ST',
+  'SN',
+  'SC',
+  'SL',
+  'SO',
+  'ZA',
+  'SS',
+  'TZ',
+  'TG',
+  'UG',
+  'ZM',
+  'ZW'].join(';');
