+++ src/redux/co2Emissions/co2Emissions.js
@@ -1,25 +1,25 @@
-import fetchCo2Data from '../api/handler';
+import fetchCo2Data from './api/handler';
 
 const FETCH_CO2EMISSIONS = 'FETCH_CO2EMISSIONS';
 
 const initialState = [];
 
 export const getCo2Data = (payload) => ({
-    type: FETCH_CO2EMISSIONS,
-    payload,
+  type: FETCH_CO2EMISSIONS,
+  payload,
 });
 
 export const getCo2DataApi = (country) => async (dispatch) => {
-    await fetchCo2Data(country).then((data) => dispatch(getCo2Data(data)));
-}
+  await fetchCo2Data(country).then((data) => dispatch(getCo2Data(data)));
+};
 
 const reducer = (state = initialState, action) => {
-    switch (action.type) {
-        case FETCH_CO2EMISSIONS:
-            return action.payload;
-        default:
-            return state;
-    }
-}
+  switch (action.type) {
+    case FETCH_CO2EMISSIONS:
+      return action.payload;
+    default:
+      return state;
+  }
+};
 
 export default reducer;
